requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
gunicorn>=20.1.0
uvicorn[standard]>=0.25.0
dj-database-url>=2.1.0
whitenoise>=6.6.0
boto3>=1.34.0
django-storages>=1.14.0
channels_redis>=4.2.0
